backend/requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1

# Database & Migrations
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # Required for PostgreSQL

# Security & Authentication
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7
python-dotenv==1.0.0

# API & Coding Engine
requests==2.31.0       # For Judge0 API calls
python-dateutil==2.8.2

# Production Server (Used by Docker)
gunicorn==21.2.0

Flask-Migrate==4.0.5

Flask-Limiter

requests

# run inside the terminal 
# pip install -r requirements.txt
# docker-compose up --build